requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
packaging==24.2
pillow==11.0.0
propcache==0.2.0
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
requests==2.32.3
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
